requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
requests
python-dotenv
ipython